Refactors sales analyst into merchant and item analysts.
