+++ lib/sales_engine.rb
@@ -1,6 +1,7 @@
 require_relative 'merchant_repository'
 require_relative 'item_repository'
 require 'csv'
+require 'time'
 
 class SalesEngine
